# backend/src/main/java/com/itacademy/dto/AuthorDto.java
package com.itacademy.dto;

import lombok.AllArgsConstructor;
import lombok.Data;
import lombok.NoArgsConstructor;

@NoArgsConstructor
@AllArgsConstructor
@Data
public class AuthorDto {

    private String name;
    private String firstName;
    private String lastName;
}