# backend/src/main/java/com/itacademy/dto/RoleDto.java
package com.itacademy.dto;

import lombok.AllArgsConstructor;
import lombok.Data;
import lombok.NoArgsConstructor;

@NoArgsConstructor
@AllArgsConstructor
@Data
public class RoleDto {
//    private Integer id;
    private String name;


}